Integrating backend and frontend
- Implementing a single server instance sharing

- Implementing API to clear previous model selection state
